Conflicts:

	wafadmin/Tools/python.py

+++ src/plugins/flac/flac.c
@@ -1,5 +1,5 @@
 /*  XMMS2 - X Music Multiplexer System
- *  Copyright (C) 2003-2007 XMMS2 Team
+ *  Copyright (C) 2003-2008 XMMS2 Team
  *
  *  PLUGINS ARE NOT CONSIDERED TO BE DERIVED WORK !!!
  *
@@ -248,7 +248,7 @@ flac_callback_metadata (const FLAC__StreamDecoder *flacdecoder,
 			data->total_samples = metadata->data.stream_info.total_samples;
 
 			if (filesize > 0 && data->total_samples) {
-				data->bit_rate = (guint) (filesize * 8 *
+				data->bit_rate = (guint) ((guint64) filesize * 8 *
 				                 (guint64) data->sample_rate /
 				                 (guint64) data->total_samples);
 			}
